修改ResponseHelper返回格式，完成RIFD订单操作，完成订单管理相关接口

+++ YOY.Model/DB/Commodity.cs
@@ -50,7 +50,7 @@ namespace YOY.Model.DB
         /// </summary>
         [Column("CommodityPrice")]
         [DataMember]
-        public float CommodityPrice { get; set; }
+        public double CommodityPrice { get; set; }
 
         /// <summary>
         /// 商品描述
